+++ src/materials/MaterialOutputFlags.ts
@@ -1,0 +1,16 @@
+//
+// Allows a shader to change its output with recompilation
+//
+// Authors:
+// * @bhouston
+//
+
+export enum MaterialOutputFlags {
+  Beauty = 0x1,
+  Albedo = 0x2,
+  Normal = 0x4,
+  Depth = 0x8,
+  Roughness = 0x10,
+  Metalness = 0x20,
+  Emissive = 0x40,
+}