+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-All notable changes to this project will be documented in this file for each PR. 
+All notable changes to this project will be documented in this file for each PR.
 Ideally, a Github Action will vet changes to this file for a PR to be validated.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-### Added 
+### Added
+
 - Maintainance of a changelog.
 - Added glsl transpiler to JS modules.
+- Added VirtualFramebuffer, CanvasFramebuffer.
+- Added MaterialOutputFlags.
+- Rename boxGeometry to box.
+- Add count and offset to VertexAttributeGeometry, VertexArrayObject.